+++ Modules/Ruleset.Firewall/Public/Format-Output.ps1
@@ -53,7 +53,7 @@ PS> Net-NewFirewallRule ... | Format-Output
 None. Format-Output does not generate any output
 
 .NOTES
-None.
+TODO: For force loaded rules it should say: "Force Load Rule:"
 #>
 function Format-Output
 {