+++ Readme/CHANGELOG.md
@@ -9,6 +9,7 @@ If you would like to see fresh changes done since last release you can do so on 
 
 - [Changelog](#changelog)
   - [Table of contents](#table-of-contents)
+  - [v0.9.1 (develop branch)](#v091-develop-branch)
   - [v0.9.0 (current release)](#v090-current-release)
   - [v0.8.0](#v080)
   - [v0.7.0](#v070)
@@ -18,6 +19,14 @@ If you would like to see fresh changes done since last release you can do so on 
   - [v0.4.0](#v040)
   - [v0.3.0](#v030)
   - [v0.2](#v02)
+
+## v0.9.1 (develop branch)
+
+**NOTE:** Changes for unreleased version (develop branch) may change or be upfront
+
+- Rules
+
+  - Updated rule script for git and GitHub Desktop
 
 ## v0.9.0 (current release)